+++ src/ccbr_actions/actions.py
@@ -8,6 +8,7 @@ import requests
 import uuid
 import warnings
 
+from .github import GITHUB_API_URL, github_api_headers, github_api_post
 from .util import path_resolve
 from .versions import get_latest_release_tag
 
@@ -108,20 +109,14 @@ def trigger_workflow(workflow_name, branch, repo, inputs=None, debug=False):
         branch (str): The branch to trigger the workflow on.
         repo (str): The GitHub repository to trigger the workflow in.
     """
-    url = f"https://api.github.com/repos/{repo}/actions/workflows/{workflow_name}/dispatches"
-    headers = {
-        "Accept": "application/vnd.github.v3+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-    }
-    if (
-        "GITHUB_TOKEN" in os.environ
-    ):  # required for curl usage, not needed for gh api usage
-        headers["Authorization"] = f"token {os.environ['GITHUB_TOKEN']}"
+    url = f"{GITHUB_API_URL}/repos/{repo}/actions/workflows/{workflow_name}/dispatches"
+    token = os.environ.get("GITHUB_TOKEN")
+    headers = github_api_headers(token=token)
     data = {"ref": branch}
     if inputs and isinstance(inputs, dict):
         data.update(inputs)
     if not debug:
-        response = requests.post(url, headers=headers, json=data)
+        response = github_api_post(url=url, token=token, json=data)
         if response.status_code != 204:
             warnings.warn(f"Failed to trigger workflow:\n{response.text}")
         return response