+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## actions development version
 
 - New action: `changed-files` to detect files that changed matching a pattern. (#152, @kelly-sovacool)
+  - Can compare the most recent commit only or the entire PR changes by setting the `comparison-mode` option. (#156, @kelly-sovacool)
 - Bump actions versions to the latest release in anticipation of the node.js 20 deprecation. (#149, @kelly-sovacool)
 - Improve `maintain-milestones`:
   - Only close the immediately prior year's milestones after May of the current year for performance reviews. (#145, @kelly-sovacool)