+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## actions development version
+## actions 0.6.0
 
 - New action: `changed-files` to detect files that changed matching a pattern. (#152, @kelly-sovacool)
   - Can compare the most recent commit only or the entire PR changes by setting the `comparison-mode` option. (#156, @kelly-sovacool)